# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        description: Para erros de validação
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  handlers.AddLabTestToTemplateRequest:
    properties:
      testId:
        type: string
    required:
    - testId
    type: object
  handlers.UpdateLabRequestTemplateTestsRequest:
    properties:
      testIds:
        items:
          type: string
        type: array
    required:
    - testIds
    type: object
  models.Article:
    properties:
      abstract:
        description: Conteúdo
        type: string
      articleType:
        description: |-
          Classificação
          @enum research_article,review,meta_analysis,case_study,clinical_trial,editorial,letter,protocol,lecture
        enum:
        - research_article
        - review
        - meta_analysis
        - case_study
        - clinical_trial
        - editorial
        - letter
        - protocol
        - lecture
        type: string
      authors:
        description: |-
          @minLength 1
          @maxLength 2000
        maxLength: 2000
        minLength: 1
        type: string
      createdAt:
        description: Timestamps
        type: string
      createdBy:
        description: Audit
        type: string
      doi:
        description: |-
          Identificadores Únicos
          @example 10.1038/s41586-024-07146-0
          @maxLength 255
        maxLength: 255
        type: string
      favorite:
        description: Gestão Pessoal
        type: boolean
      fileHash:
        description: |-
          Arquivo
          @example a1b2c3d4e5f6...
          @maxLength 64
        type: string
      fileSize:
        description: '@example 2457600'
        minimum: 0
        type: integer
      fullContent:
        type: string
      id:
        description: |-
          Primary Key
          @example 550e8400-e29b-41d4-a716-446655440000
        type: string
      indexedAt:
        description: Metadados do Sistema
        type: string
      internalLink:
        description: '@example /uploads/articles/550e8400-e29b-41d4-a716-446655440000.pdf'
        maxLength: 2048
        type: string
      issn:
        description: |-
          @example 0028-0836
          @maxLength 20
        maxLength: 20
        type: string
      journal:
        description: |-
          @minLength 1
          @maxLength 500
        maxLength: 500
        minLength: 1
        type: string
      keywords:
        description: '@items.type string'
        items:
          type: string
        type: array
      language:
        description: |-
          @minLength 2
          @maxLength 10
          @example en
        maxLength: 10
        minLength: 2
        type: string
      lastAccessedAt:
        type: string
      meshTerms:
        description: '@items.type string'
        items:
          type: string
        type: array
      notes:
        description: Notas pessoais do médico
        type: string
      originalLink:
        description: |-
          Links
          @example https://www.nature.com/articles/s41586-024-07146-0
        maxLength: 2048
        type: string
      pmid:
        description: |-
          @example 38355808
          @maxLength 50
        maxLength: 50
        type: string
      publishDate:
        description: '@example 2024-03-15'
        type: string
      rating:
        description: |-
          @minimum 0
          @maximum 5
          @example 4
        maximum: 5
        minimum: 0
        type: integer
      scoreItems:
        description: |-
          Relationships
          @items.type object
        items:
          $ref: '#/definitions/models.ScoreItem'
        type: array
      specialty:
        description: |-
          @example Cardiology
          @maxLength 200
        maxLength: 200
        type: string
      title:
        description: |-
          Metadados Bibliográficos Essenciais
          @minLength 3
          @maxLength 1000
        maxLength: 1000
        minLength: 3
        type: string
      updatedAt:
        type: string
      updatedBy:
        description: User ID que atualizou
        type: string
    required:
    - articleType
    - authors
    - journal
    - language
    - publishDate
    - title
    type: object
  models.Gender:
    enum:
    - male
    - female
    - other
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
    - GenderOther
...
  models.LabRequestTemplate:
    description: Template pré-configurado com lista de exames para facilitar solicitações
    properties:
      createdAt:
        description: Timestamps
        type: string
      description:
        description: |-
          Descrição do template
          @example "Exames de rotina para check-up anual geral"
        type: string
      displayOrder:
        description: Ordem de exibição/prioridade
        type: integer
      id:
        description: ID único do template
        type: string
      isActive:
        description: Status (ativo/inativo)
        type: boolean
      labTests:
        description: Exames incluídos neste template (many-to-many)
        items:
          $ref: '#/definitions/models.LabTestDefinition'
        type: array
      name:
        description: |-
          Nome do template
          @example "Check-up Anual", "Avaliação Cardíaca", "Perfil Tireoidiano"
          @required
        maxLength: 200
        type: string
      updatedAt:
        type: string
    required:
    - name
    type: object
...
  models.LabTestCategory:
    enum:
    - hematology
    - biochemistry
    - hormones
    - immunology
    - microbiology
    - urine
    - imaging
    - functional
    - genetics
    - other
    type: string
    x-enum-comments:
      LabTestCategoryBiochemistry: Glicose, ureia, creatinina
      LabTestCategoryFunctional: Medicina funcional
      LabTestCategoryGenetics: Testes genéticos
      LabTestCategoryHematology: Hemograma, coagulação
      LabTestCategoryHormones: TSH, T4, testosterona
      LabTestCategoryImaging: Raio-X, TC, RM
      LabTestCategoryImmunology: Sorologias, autoimunes
      LabTestCategoryMicrobiology: Culturas, antibiogramas
      LabTestCategoryOther: Outros
      LabTestCategoryUrine: EAS, urocultura
    x-enum-descriptions:
    - Hemograma, coagulação
    - Glicose, ureia, creatinina
    - TSH, T4, testosterona
    - Sorologias, autoimunes
    - Culturas, antibiogramas
    - EAS, urocultura
    - Raio-X, TC, RM
    - Medicina funcional
    - Testes genéticos
    - Outros
    x-enum-varnames:
    - LabTestCategoryHematology
    - LabTestCategoryBiochemistry
    - LabTestCategoryHormones
    - LabTestCategoryImmunology
    - LabTestCategoryMicrobiology
    - LabTestCategoryUrine
    - LabTestCategoryImaging
    - LabTestCategoryFunctional
    - LabTestCategoryGenetics
    - LabTestCategoryOther
  models.LabTestDefinition:
    description: Definição estruturada de exames e seus parâmetros
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.LabTestCategory'
        description: |-
          Categoria do exame
          @enum hematology,biochemistry,hormones,immunology,microbiology,urine,imaging,functional,genetics,other
      clinicalIndications:
        description: Indicações clínicas principais
        type: string
      clinicalRecommendations:
        description: |-
          Recomendações clínicas (150-300 palavras)
          Quando solicitar, interpretação de resultados, fatores que afetam valores, intervenções
        type: string
      clinicalSignificance:
        description: |-
          Significância clínica detalhada (200-400 palavras)
          Mecanismos fisiológicos, aplicações clínicas, interpretação de valores alterados
        type: string
      code:
        description: |-
          Código interno único (usado para referência programática)
          @example HEMOGRAMA_COMPLETO, HGB, GLUCOSE_FASTING
        maxLength: 100
        type: string
      collectionMethod:
        description: |-
          Método de coleta/realização
          @example Sangue venoso com jejum de 8-12h
        type: string
      createdAt:
        description: Timestamps
        type: string
      description:
        description: Descrição/observações sobre o exame
        type: string
      displayOrder:
        description: Ordem de exibição (para organizar parâmetros de um exame)
        type: integer
      fastingHours:
        description: Tempo de jejum necessário (em horas)
        type: integer
      id:
        description: ID único da definição
        type: string
      isActive:
        description: Status (ativo/inativo)
        type: boolean
      isRequestable:
        description: |-
          Indica se o exame pode ser solicitado individualmente
          true: pode ser solicitado (ex: Hemograma Completo, Glicemia)
          false: só aparece como resultado de outro exame (ex: Hemoglobina, Bilirrubina Indireta)
        type: boolean
      loincCode:
        description: |-
          Código LOINC (Logical Observation Identifiers Names and Codes)
          Padrão internacional para identificação de exames
          @example 718-7 (Hemoglobin)
        type: string
      longevityContext:
        description: |-
          Contexto de longevidade (100-200 palavras)
          Relação com envelhecimento saudável, marcadores de longevidade, implicações preventivas
        type: string
      name:
        description: |-
          Nome do exame/parâmetro
          @example Hemograma Completo, Hemoglobina, Glicemia de Jejum
        maxLength: 300
        type: string
      parentTest:
        allOf:
        - $ref: '#/definitions/models.LabTestDefinition'
        description: Exame pai (se for um parâmetro/subexame)
      parentTestId:
        description: |-
          ID do exame pai (hierarquia)
          Ex: Hemoglobina tem parentTestId = Hemograma Completo
          Ex: Bilirrubina Indireta tem parentTestId = Bilirrubina Total e Frações
        type: string
      resultType:
        description: |-
          Tipo de resultado
          @enum numeric, text, boolean, categorical
        type: string
      scoreMappings:
        description: Mapeamentos para itens do escore
        items:
          $ref: '#/definitions/models.LabTestScoreMapping'
        type: array
      shortName:
        description: |-
          Nome curto/abreviação (opcional)
          @example Hemograma, Hb, Gli
        type: string
      specimenType:
        description: |-
          Material biológico
          @example Sangue total, Soro, Urina, Fezes
        type: string
      subTests:
        description: Subexames/parâmetros (se for um exame composto)
        items:
          $ref: '#/definitions/models.LabTestDefinition'
        type: array
      tussCode:
        description: |-
          Código TUSS (Terminologia Unificada da Saúde Suplementar)
          Usado para faturamento e solicitação no Brasil
          @example 40304485
        type: string
      unit:
        description: |-
          Unidade de medida padrão
          @example g/dL, mg/dL, mU/L, %
        type: string
      unitConversion:
        description: |-
          Fórmula de conversão entre unidades (se aplicável)
          @example 1 g/dL = 10 g/L
        type: string
      updatedAt:
        type: string
    required:
    - category
    - code
    - name
    - resultType
    type: object
  models.LabTestScoreMapping:
    description: Mapeamento entre exame e item do escore Plenya
    properties:
      createdAt:
        description: Timestamps
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/models.Gender'
        description: |-
          Especificação de gênero (se o mapeamento for específico)
          @enum male, female, null (ambos)
      id:
        type: string
      isActive:
        description: Status
        type: boolean
      labTest:
        allOf:
        - $ref: '#/definitions/models.LabTestDefinition'
        description: Relacionamentos
      labTestId:
        description: ID do exame laboratorial
        type: string
      maxAge:
        description: Faixa etária máxima (em anos)
        type: integer
      minAge:
        description: Faixa etária mínima (em anos)
        type: integer
      notes:
        description: Observações sobre o mapeamento
        type: string
      scoreItem:
        $ref: '#/definitions/models.ScoreItem'
      scoreItemId:
        description: ID do item do escore
        type: string
      updatedAt:
        type: string
    required:
    - labTestId
    - scoreItemId
    type: object
...
  models.ScoreItem:
    description: Item de escore - parâmetro clínico específico
    properties:
      articles:
        description: |-
          Many-to-many relationship with Articles
          @items.type object
        items:
          $ref: '#/definitions/models.Article'
        type: array
      childItems:
        items:
          $ref: '#/definitions/models.ScoreItem'
        type: array
      clinicalRelevance:
        description: |-
          Relevância clínica - explicação técnica para profissionais de saúde
          @example Valores baixos de hemoglobina indicam anemia, que pode estar associada a fadiga, redução da capacidade funcional e aumento do risco cardiovascular
        type: string
      code:
        description: |-
          Código único para identificação programática (opcional)
          @example HGB_M, GLUCOSE_FASTING, IMC, BREASTFEEDING_DURATION
        maxLength: 100
        type: string
      conduct:
        description: |-
          Conduta clínica recomendada
          @example Investigar causa da anemia (deficiência de ferro, B12, folato, doença crônica). Suplementação conforme indicação. Encaminhar ao hematologista se Hb < 10 g/dL ou causa não esclarecida
        type: string
      createdAt:
        description: Timestamps
        type: string
      id:
        description: '@example 550e8400-e29b-41d4-a716-446655440000'
        type: string
      lastReview:
        description: |-
          Data da última revisão dos campos clínicos ou artigos associados
          @example 2026-01-25T10:30:00Z
        type: string
      levels:
        items:
          $ref: '#/definitions/models.ScoreLevel'
        type: array
      name:
        description: |-
          @minLength 2
          @maxLength 300
          @example Hemoglobina - Homens
        maxLength: 300
        minLength: 2
        type: string
      order:
        description: |-
          @minimum 0
          @maximum 9999
          @example 1
        maximum: 9999
        minimum: 0
        type: integer
      parentItem:
        $ref: '#/definitions/models.ScoreItem'
      parentItemId:
        description: |-
          Self-referencing for hierarchical items (optional)
          @example 550e8400-e29b-41d4-a716-446655440000
        type: string
      patientExplanation:
        description: |-
          Explicação para o paciente - linguagem simples e acessível
          @example Hemoglobina é a proteína que transporta oxigênio no sangue. Valores baixos podem causar cansaço e falta de ar
        type: string
      points:
        description: |-
          @minimum 0
          @maximum 100
          @example 20
        maximum: 100
        minimum: 0
        type: number
      subgroup:
        allOf:
        - $ref: '#/definitions/models.ScoreSubgroup'
        description: Relationships
      subgroupId:
        description: |-
          Foreign Keys
          @example 550e8400-e29b-41d4-a716-446655440000
        type: string
      unit:
        description: '@example g/dL'
        maxLength: 50
        type: string
      unitConversion:
        description: '@example 1 g/dL = 10 g/L'
        type: string
      updatedAt:
        type: string
    required:
    - name
    - subgroupId
    type: object
...
host: localhost:3001
info:
  contact:
    email: suporte@plenya.com.br
    name: API Support
  description: Sistema de Prontuário Médico Eletrônico
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Plenya EMR API
  version: "1.0"
paths:
  /api/v1/lab-request-templates:
    get:
      parameters:
      - default: false
        description: Include lab tests
        in: query
        name: withTests
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LabRequestTemplate'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all lab request templates
      tags:
      - LabRequestTemplates
    post:
      consumes:
      - application/json
      parameters:
      - description: Lab request template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.LabRequestTemplate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create lab request template
      tags:
      - LabRequestTemplates
  /api/v1/lab-request-templates/{id}:
    delete:
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete lab request template
      tags:
      - LabRequestTemplates
    get:
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get lab request template by ID
      tags:
      - LabRequestTemplates
    put:
      consumes:
      - application/json
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated lab request template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/models.LabRequestTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update lab request template
      tags:
      - LabRequestTemplates
  /api/v1/lab-request-templates/{id}/tests:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      - description: Test ID
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/handlers.AddLabTestToTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Add lab test to template
      tags:
      - LabRequestTemplates
    put:
      consumes:
      - application/json
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      - description: Test IDs
        in: body
        name: tests
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLabRequestTemplateTestsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update lab request template tests
      tags:
      - LabRequestTemplates
  /api/v1/lab-request-templates/{id}/tests/{testId}:
    delete:
      parameters:
      - description: Lab request template ID
        in: path
        name: id
        required: true
        type: string
      - description: Test ID to remove
        in: path
        name: testId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LabRequestTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Remove lab test from template
      tags:
      - LabRequestTemplates
  /api/v1/lab-request-templates/search:
    get:
      parameters:
      - description: Search term
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LabRequestTemplate'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Search lab request templates
      tags:
      - LabRequestTemplates
...
securityDefinitions:
  BearerAuth:
    description: Digite 'Bearer' seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"